% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\name{binIons}
\alias{binIons}
\title{Create profile data from ions}
\usage{
binIons(ions, fun = max, mzBins = 500, rtBins = mzBins)
}
\description{
This function takes a matrix of ion triples and bins the data to create well
formated data
}

