% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsIonList.R
\docType{methods}
\name{chroms,MsIonList-method}
\alias{chroms,MsIonList-method}
\title{Extract chromatograms from an ion list}
\usage{
\S4method{chroms}{MsIonList}(object, ...)
}
\description{
Extract chromatograms from an ion list
}

