% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsPeakList.R
\docType{methods}
\name{chroms,MsPeakList-method}
\alias{chroms,MsPeakList-method}
\title{Get chroms with optional expansion}
\usage{
\S4method{chroms}{MsPeakList}(object, mzExpand = 0, rtExpand = 0)
}
\description{
Get chroms with optional expansion
}

