% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsScanList.R
\docType{methods}
\name{chroms,MsScanList-method}
\alias{chroms,MsScanList-method}
\title{Get chroms with optional expansion}
\usage{
\S4method{chroms}{MsScanList}(object, mz, expand = 0)
}
\description{
Get chroms with optional expansion
}

