% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chroms}
\alias{chroms}
\title{Extract chromatograms from an object}
\usage{
chroms(object, ...)
}
\arguments{
\item{object}{The object to extract chromatograms from}

\item{...}{Aditional parameters depending on the class of @object}
}
\value{
An MsChromList object
}
\description{
This function allows the user to extract chromatograms from a range of
objects intoan MsChromList object
}

