% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsData.R
\name{createMsData}
\alias{createMsData}
\title{Create an MsData object from a raw MS data file}
\usage{
createMsData(rawFile, saryName, force = FALSE)
}
\arguments{
\item{rawFile}{The path to an mzR-compliant MS data file}
}
\value{
An MsData object
}
\description{
This function takes an mzR-compliant data file and create the initial,
minimal sary database file and set up the link to both the database and raw
file.
}
\examples{
file <- 'test'
msdata <- createMsData(file)
}
\seealso{
\code{\link{loadMsData}} \code{\linkS4class{MsData}}
}

