% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsDataSet.R
\name{createMsDataSet}
\alias{createMsDataSet}
\title{Create an MsDataSet from a list of raw files}
\usage{
createMsDataSet(rawFiles, setName, memberNames, ref, design, force = FALSE)
}
\description{
Create an MsDataSet from a list of raw files
}

