% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsData.R
\docType{methods}
\name{detectPeaks,MsData-method}
\alias{detectPeaks,MsData-method}
\title{Peak detection in MsData object}
\usage{
\S4method{detectPeaks}{MsData}(object, method, ...)
}
\description{
Peak detection in MsData object
}

