% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{editHistory}
\alias{editHistory}
\title{Get a list of all edits on an object}
\usage{
editHistory(object)
}
\arguments{
\item{object}{An object containing an MsConnections object}
}
\value{
A data.frame with edits from first to last
}
\description{
This function returns the edit history as stored in the history table in the
sary database.
}

