% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expand.call}
\alias{expand.call}
\title{Return a call in which all of the arguments which were supplied
or have presets are specified by their full names and supplied
or default values.}
\usage{
expand.call(definition = NULL, call = sys.call(sys.parent(1)),
  expand.dots = TRUE, doEval = TRUE)
}
\arguments{
\item{definition}{a function. See \code{\link[base]{match.call}}.}

\item{call}{an unevaluated call to the function specified by definition.
See \code{\link[base]{match.call}}.}

\item{expand.dots}{logical. Should arguments matching ... in the call be
included or left as a ... argument? See \code{\link[base]{match.call}}.}

\item{doEval}{logical, defaults to TRUE. Should function arguments be
 evaluated in the returned call or not?}
}
\value{
An object of class call.
}
\description{
Return a call in which all of the arguments which were supplied
or have presets are specified by their full names and supplied
or default values.
}
\author{
fabians
}
\references{
\href{http://stackoverflow.com/questions/3478923/displaying-the-actual-parameter-list-of-the-function-during-execution}{Stack Overflow answer}
}
\seealso{
\code{\link[base]{match.call}}
}

