% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsChromList.R
\docType{methods}
\name{ions,MsChromList-method}
\alias{ions,MsChromList-method}
\title{Extract ions from the EIC}
\usage{
\S4method{ions}{MsChromList}(object, ...)
}
\description{
Extract ions from the EIC
}

