% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ions,MsData-method}
\alias{ions,MsData-method}
\title{Extract ions from an MsData object}
\usage{
\S4method{ions}{MsData}(object, seqNum, retentionTime, mzRange, msLevel = 1,
  SIMPLIFY = TRUE)
}
\description{
Extract ions from an MsData object
}

