% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsPeakList.R
\docType{methods}
\name{ions,MsPeakList-method}
\alias{ions,MsPeakList-method}
\title{Get ions with optional expansion}
\usage{
\S4method{ions}{MsPeakList}(object, mzExpand = 0, rtExpand = 0, ...)
}
\description{
Get ions with optional expansion
}

