% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsScanList.R
\docType{methods}
\name{ions,MsScanList-method}
\alias{ions,MsScanList-method}
\title{Get ions with optional expansion}
\usage{
\S4method{ions}{MsScanList}(object, mz, expand = 0, ...)
}
\description{
Get ions with optional expansion
}

