% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ions}
\alias{ions}
\title{Extract ions from an object}
\usage{
ions(object, ...)
}
\arguments{
\item{object}{The object to extract ions from}

\item{...}{Aditional parameters depending on the class of @object}
}
\value{
An MsIonList object
}
\description{
This function allows the user to extract ions from a range of
objects into an MsIonList object
}

