% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isEmpty}
\alias{isEmpty}
\title{Test whether the elements of an object are empty}
\usage{
isEmpty(object)
}
\arguments{
\item{object}{A list-like object from MSsary}
}
\value{
A logical vector of the same length as object
}
\description{
This function applies to list-like objects such as MsScanList and tests
whether the individual elements of the object are empty (i.e. doesn't contain
any ions). It returns a logical vector of the same length as the object. If
\code{TRUE} the corresponding element is empty
}

