% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isRaw}
\alias{isRaw}
\title{Test whether the elements of an object are based on raw value}
\usage{
isRaw(object)
}
\arguments{
\item{object}{A list-like object from MSsary}
}
\value{
A logical vector of the same length as object
}
\description{
This function applies to list-like objects such as MsScanList and tests
whether the individual elements of the object are based on raw unaltered
values or modified ones. Note that this function just test for strictly raw-
ness in the sense that it tests whether \code{raw=TRUE} was set during
construction. If the raw data has not been modified it is possible to have
raw and unraw MsLists that are numerically identical.
}

