% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadMsData}
\alias{loadMsData}
\title{Load in an already created sary file as an MsData object}
\usage{
loadMsData(saryFile, rawFile, testIntegrity = TRUE, testSize = 10)
}
\arguments{
\item{saryFile}{The path to the sary database file}

\item{rawFile}{The path to the raw datafile}

\item{testIntegrity}{Should the raw and sary file be compared to each other}

\item{testSize}{The number of randomly selected scans to use for comparing
raw and sary if \code{testIntegrity=TRUE}}
}
\value{
An MsData object
}
\description{
This function creates a new MsData object from an already created sary file,
together with the accompanying raw data file. The function automatically
checks the sary file and see if it complies with the format. By default it
also does a light check on whether the data in the raw file corresponds to
the data in the sary file.
}
\seealso{
\code{\link{createMsData}} \code{\linkS4class{MsData}}
}

