% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsChromList.R, R/MsIonList.R, R/MsPeakList.R, R/MsScanList.R, R/generics.R
\docType{methods}
\name{meltMS,MsChromList-method}
\alias{meltMS}
\alias{meltMS,MsChromList-method}
\alias{meltMS,MsIonList-method}
\alias{meltMS,MsPeakList-method}
\alias{meltMS,MsScanList-method}
\title{Create ggplot2 ready data from object}
\usage{
\S4method{meltMS}{MsChromList}(object)

\S4method{meltMS}{MsIonList}(object, extraData)

\S4method{meltMS}{MsPeakList}(object)

\S4method{meltMS}{MsScanList}(object, extraData)

meltMS(object, ...)
}
\arguments{
\item{object}{An object to be converted to a data.frame}
}
\value{
A data.frame with the relevant information
}
\description{
This is an internal function that converts objects to long-format data.frames
ready for use in ggplot2 functions.
}
\section{Methods (by class)}{
\itemize{
\item \code{MsChromList}: melt an MsChromList

\item \code{MsIonList}: melt an MsIonList

\item \code{MsPeakList}: melt an MsPeakList

\item \code{MsScanList}: melt an MsScanList
}}
\keyword{internal}

