% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MsChromList-class}
\alias{MsChromList-class}
\alias{[,MsChromList,logical,missing,missing-method}
\alias{[,MsChromList,numeric,missing,missing-method}
\alias{msData,MsChromList-method}
\alias{msInfo,MsChromList-method}
\alias{msPlot,MsChromList-method}
\alias{names,MsChromList-method}
\alias{show,MsChromList-method}
\title{Class to handle extracted ion chromatograms}
\usage{
\S4method{show}{MsChromList}(object)

\S4method{names}{MsChromList}(x)

\S4method{[}{MsChromList,numeric,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{MsChromList,logical,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{msPlot}{MsChromList}(object, type = c("TIC", "BPC"),
  collapse = FALSE, ...)

\S4method{msInfo}{MsChromList}(object)

\S4method{msData}{MsChromList}(object)
}
\arguments{
\item{object}{An MsChromList object}

\item{x}{An MsChromList object}

\item{type}{Which type of plots should be plotted}

\item{collapse}{Should 'BPC' and 'TIC' be overlayed or on separate plots,
TRUE means overlayed.}

\item{x}{An MsChromList object}
}
\value{
\code{msInfo}: A data.frame with one row containing information
about the chromatograms

\code{msData}: A list of matrices with acquisitionNum, retentionTime,
totIonCurrent and basePeakIntensity for each scan in the chromatogram
}
\description{
This class is a container for one or several XIC from MsData object(s).
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of object

\item \code{names}: Get the names of the scans

\item \code{[}: Subset an MsList object

\item \code{[}: Subset an MsList object

\item \code{msPlot}: Create a plot

\item \code{msInfo}: Get information about the chromatograms

\item \code{msData}: Get the chromatogram data
}}
\section{Slots}{

\describe{
\item{\code{connections}}{A list of MsConnections}

\item{\code{info}}{A data.frame with information on the contained XIC}

\item{\code{data}}{A matrix containing the raw data pertaining to the XIC}

\item{\code{mapping}}{A matrix with the mapping of XIC to the rows in the
@data matrix}
}}
\seealso{
Other MSsary-classees: \code{\link{MsIonList-class}},
  \code{\link{msData,MsIonList-method}},
  \code{\link{msInfo,MsIonList-method}},
  \code{\link{msPlot,MsIonList-method}},
  \code{\link{names,MsIonList-method}},
  \code{\link{show,MsIonList-method}};
  \code{\link{MsPeakList-class}},
  \code{\link{[,MsPeakList,logical,missing,missing-method}},
  \code{\link{[,MsPeakList,numeric,missing,missing-method}},
  \code{\link{msData,MsPeakList-method}},
  \code{\link{msInfo,MsPeakList-method}},
  \code{\link{msPlot,MsPeakList-method}},
  \code{\link{names,MsPeakList-method}},
  \code{\link{show,MsPeakList-method}};
  \code{\link{MsScanList-class}},
  \code{\link{[,MsScanList,logical,missing,missing-method}},
  \code{\link{[,MsScanList,numeric,missing,missing-method}},
  \code{\link{children,MsScanList-method}},
  \code{\link{msData,MsScanList-method}},
  \code{\link{msInfo,MsScanList-method}},
  \code{\link{msPlot,MsScanList-method}},
  \code{\link{names,MsScanList-method}},
  \code{\link{nextScan,MsScanList-method}},
  \code{\link{parent,MsScanList-method}},
  \code{\link{previousScan,MsScanList-method}},
  \code{\link{scanMode,MsScanList-method}},
  \code{\link{show,MsScanList-method}},
  \code{\link{siblings,MsScanList-method}}
}

