% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MsConnections-class}
\alias{MsConnections}
\alias{MsConnections-class}
\title{A class to handle MS I/O}
\description{
This class takes care of the communication with the sary database and the raw
datafile. Its main responsibility is to ensure a stable connection across
sessions without causing the session to crash (This would normally happen if
an mzRramp object was accessed in a different session than the one it is
created in). It relies on knowledge of the location of the raw and database
files and will throw an error if these files are moved.
}
\details{
TODO: Schema for updating file location vs. creating a new object (which is
cheap). Decide whether this class also should handle writing to the sary db.
}
\section{Fields}{

\describe{
\item{\code{rawFile}}{The path to the location of the MS raw data file}

\item{\code{ramp}}{Object of class mzRramp. A connection to the MS raw data file}

\item{\code{saryFile}}{The path to the location of the sary database file}

\item{\code{saryDB}}{An SQLiteConnection object. A connection to the sary database}
}}
\section{Methods}{

\describe{
\item{\code{addData(name, data)}}{Adds data to an already created table. 'name' specifies the name of the
table, 'data' is a data.frame containing the data to be added.}

\item{\code{addTable(name, definition)}}{Creates the table 'name' with the columns defined in 'definition'.
The definition can be a character vector that will be concatenated with ', '.}

\item{\code{getScans(ids, raw = FALSE)}}{Transparently extract scans, getting modified scans if they exists 
or read from the raw file if they don't}

\item{\code{raw()}}{Get the mzRramp connection in a safe manner}

\item{\code{sary()}}{Get the sary database connection in a safe manner}

\item{\code{setScans(ids, scans)}}{Set scans to new values}

\item{\code{validate(subset = 10)}}{Checks that the header data matches between raw and sary}

\item{\code{verify()}}{Verify that the default tables are present and their structure conforms}
}}
\seealso{
\code{\linkS4class{MsData}}
}

