% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsData.R
\docType{class}
\name{MsData-class}
\alias{MsData-class}
\alias{editHistory,MsData-method}
\alias{length,MsData-method}
\alias{show,MsData-method}
\title{Main class for interacting with MS data}
\usage{
\S4method{show}{MsData}(object)

\S4method{length}{MsData}(x)

\S4method{editHistory}{MsData}(object)
}
\arguments{
\item{object}{An MsData object}

\item{x}{An MsData object}
}
\description{
This class is the main entry point for interacting with MS data in the MSsary
package. It is designed to seamlessly integrate data from raw data files and
additional results from various algorithms. Furthermore it provides a mean to
commit persistent changes to the raw data (e.g. filtering or recalibration),
that will be available across sessions in a non-destructive manner.
}
\details{
Behind the scene an SQLite-based file is created that handles all deviations
and additions to the raw data. The file is named after the raw data file but
has a .sary file extension to illustrate its relationship to this package.
Because it is just an SQLite database file the data generated with MSsary can
easily be accessed from other programs.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of content

\item \code{length}: Number of scans in the raw data

\item \code{editHistory}: Get the edit history of the object
}}
\section{Slots}{

\describe{
\item{\code{connections}}{An MsConnections reference class object}
}}
\section{Constructors}{

Objects of class MsData can be created either from scratch with
\code{\link{createMsData}} or from an already created sary-file
\code{\link{loadMsData}}.
}
\seealso{
\code{\link{createMsData}} \code{\link{loadMsData}}
}

