% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsDataSet.R
\docType{class}
\name{MsDataSet-class}
\alias{MsDataSet-class}
\title{Store and operate on sets of samples}
\description{
This object handles multiple MsData sets and facilitates batch analysis on
these.
}

