% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MsIonList-class}
\alias{MsIonList-class}
\alias{msData,MsIonList-method}
\alias{msInfo,MsIonList-method}
\alias{msPlot,MsIonList-method}
\alias{names,MsIonList-method}
\alias{show,MsIonList-method}
\title{Class to handle raw ion reads}
\usage{
\S4method{show}{MsIonList}(object)

\S4method{names}{MsIonList}(x)

\S4method{msPlot}{MsIonList}(object, type = "2d", simple = FALSE,
  precursors = !simple, ...)

\S4method{msInfo}{MsIonList}(object)

\S4method{msData}{MsIonList}(object)
}
\arguments{
\item{object}{An MsIonList object}

\item{x}{An MsIonList object}

\item{type}{Which type of plots should be plotted}
}
\value{
\code{msInfo}: A data.frame with one row containing information
about the ion slice

\code{msData}: A list of matrices with acquisitionNum, retentionTime,
totIonCurrent and basePeakIntensity for each scan in the chromatogram
}
\description{
This class is a container for ion data from a slice of an MsData object.
Unlike other MsList subclasses this class can only contain one slice.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of object

\item \code{names}: Get the names of the scans

\item \code{msPlot}: Create a plot

\item \code{msInfo}: Get information about the ion slice

\item \code{msData}: Get the ion data
}}
\section{Slots}{

\describe{
\item{\code{connections}}{A list of length 1 with the MsConnections object}

\item{\code{info}}{A data.frame with information on the slice}

\item{\code{data}}{A matrix containing the raw data pertaining to the ions in the
slice}

\item{\code{mapping}}{A matrix with the mapping}
}}
\seealso{
Other MSsary-classees: \code{\link{MsChromList-class}},
  \code{\link{[,MsChromList,logical,missing,missing-method}},
  \code{\link{[,MsChromList,numeric,missing,missing-method}},
  \code{\link{msData,MsChromList-method}},
  \code{\link{msInfo,MsChromList-method}},
  \code{\link{msPlot,MsChromList-method}},
  \code{\link{names,MsChromList-method}},
  \code{\link{show,MsChromList-method}};
  \code{\link{MsPeakList-class}},
  \code{\link{[,MsPeakList,logical,missing,missing-method}},
  \code{\link{[,MsPeakList,numeric,missing,missing-method}},
  \code{\link{msData,MsPeakList-method}},
  \code{\link{msInfo,MsPeakList-method}},
  \code{\link{msPlot,MsPeakList-method}},
  \code{\link{names,MsPeakList-method}},
  \code{\link{show,MsPeakList-method}};
  \code{\link{MsScanList-class}},
  \code{\link{[,MsScanList,logical,missing,missing-method}},
  \code{\link{[,MsScanList,numeric,missing,missing-method}},
  \code{\link{children,MsScanList-method}},
  \code{\link{msData,MsScanList-method}},
  \code{\link{msInfo,MsScanList-method}},
  \code{\link{msPlot,MsScanList-method}},
  \code{\link{names,MsScanList-method}},
  \code{\link{nextScan,MsScanList-method}},
  \code{\link{parent,MsScanList-method}},
  \code{\link{previousScan,MsScanList-method}},
  \code{\link{scanMode,MsScanList-method}},
  \code{\link{show,MsScanList-method}},
  \code{\link{siblings,MsScanList-method}}
}

