% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{MsList-accessors}
\alias{MsList-accessors}
\alias{msData}
\alias{msInfo}
\title{Methods for accessing data in MsList objects}
\usage{
msInfo(object)

msData(object)
}
\arguments{
\item{object}{An MsList object}
}
\value{
\code{msInfo}: A data.frame

\code{msInfo}: A list of matrices with the data for each element
}
\description{
These methods makes it possible to extract the data stored in MsList
objects.
}
\details{
\code{msInfo}: Get information about the elements

\code{msData}: Get the data stored for each element
}
\seealso{
\code{\linkS4class{MsList}}
}

