% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsList.R
\docType{class}
\name{MsList-class}
\alias{MsList-class}
\alias{[,MsList,logical,missing,missing-method}
\alias{[,MsList,numeric,missing,missing-method}
\alias{[[,MsList,numeric,missing-method}
\alias{c,MsList-method}
\alias{dropEmpty,MsList-method}
\alias{isEmpty,MsList-method}
\alias{isRaw,MsList-method}
\alias{length,MsList-method}
\alias{msData,MsList-method}
\alias{msInfo,MsList-method}
\alias{names,MsList-method}
\alias{sort,MsList-method}
\alias{uNames,MsList-method}
\title{Virtual class for MSsary list like objects}
\usage{
\S4method{length}{MsList}(x)

\S4method{names}{MsList}(x)

\S4method{uNames}{MsList}(object)

\S4method{[}{MsList,numeric,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{MsList,logical,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[[}{MsList,numeric,missing}(x, i, j)

\S4method{c}{MsList}(x, y, ..., recursive = FALSE)

\S4method{sort}{MsList}(x, decreasing = FALSE, by, ...)

\S4method{isEmpty}{MsList}(object)

\S4method{dropEmpty}{MsList}(object)

\S4method{isRaw}{MsList}(object)

\S4method{msInfo}{MsList}(object)

\S4method{msData}{MsList}(object)
}
\arguments{
\item{x}{An MsList object}
}
\description{
This class defines the storage and general methods for extracted MS data
within the MSsary framework. The class is virtual and must be sublassed for
use with e.g. scans (\code{\linkS4class{MsScanList}}). Data is stored
efficiently in a single matrix, and the class itself takes care of indexing
individual elements. Furthermore it retains a link to the MsConnections from
where the data was extracted.
}
\section{Methods (by generic)}{
\itemize{
\item \code{length}: Number of elements in the object

\item \code{names}: Get the names of the elements

\item \code{uNames}: Get unique names of elements

\item \code{[}: Subset an MsList object

\item \code{[}: Subset an MsList object

\item \code{[[}: Extract raw data

\item \code{c}: Combine multiple lists

\item \code{sort}: Sort an MsList by a property

\item \code{isEmpty}: See if elements are empty

\item \code{dropEmpty}: Removes empty elements

\item \code{isRaw}: See if elements are based on raw values

\item \code{msInfo}: Get the info for each element as a data.frame

\item \code{msData}: Get the data as a list of matrices
}}
\section{Slots}{

\describe{
\item{\code{connections}}{A list of MsConnections objects}

\item{\code{info}}{A data.frame with information pertaining to each element
(metadata)}

\item{\code{data}}{A matrix holding the actual data}

\item{\code{mapping}}{A matrix giving start and end indexes in @data for each row in
@info. Additionally it also contains the index for the MsConnections object
in the @connections list}
}}

