% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsScanList.R
\docType{class}
\name{MsScanList-class}
\alias{MsScanList-class}
\alias{[,MsScanList,logical,missing,missing-method}
\alias{[,MsScanList,numeric,missing,missing-method}
\alias{children,MsScanList-method}
\alias{msData,MsScanList-method}
\alias{msInfo,MsScanList-method}
\alias{msPlot,MsScanList-method}
\alias{names,MsScanList-method}
\alias{nextScan,MsScanList-method}
\alias{parent,MsScanList-method}
\alias{previousScan,MsScanList-method}
\alias{scanMode,MsScanList-method}
\alias{show,MsScanList-method}
\alias{siblings,MsScanList-method}
\title{Class to handle specific scans}
\usage{
\S4method{show}{MsScanList}(object)

\S4method{names}{MsScanList}(x)

\S4method{[}{MsScanList,numeric,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{MsScanList,logical,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{scanMode}{MsScanList}(object)

\S4method{msPlot}{MsScanList}(object, simple = FALSE, parent = !simple,
  precursor = !simple, peaks = !simple, masses = ifelse(simple, 0, 10),
  context = !simple, tolerance = 20, ...)

\S4method{nextScan}{MsScanList}(object, sameLevel = TRUE)

\S4method{previousScan}{MsScanList}(object, sameLevel = TRUE)

\S4method{parent}{MsScanList}(object)

\S4method{children}{MsScanList}(object)

\S4method{siblings}{MsScanList}(object)

\S4method{msInfo}{MsScanList}(object)

\S4method{msData}{MsScanList}(object)
}
\arguments{
\item{object}{An MsScanList object}

\item{x}{An MsScanList object}
}
\value{
\code{nextScan}: An MsScanList object with the next scan

\code{previousScan}: An MsScanList object with the previous scan

\code{parent}: An MsScanList object with the parent scan

\code{children}: A list of MsScanList objects with the children scans

\code{siblings}: A list of MsScanList objects with the sibling scans

\code{msInfo}: A data.frame with one row containing information
about the scan

\code{msData}: A list of matrices with m/z values in the first
column and intensity in the second
}
\description{
This class is a container for one or several scans from MsData object(s).
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of object

\item \code{names}: Get the names of the scans

\item \code{[}: Subset an MsList object

\item \code{[}: Subset an MsList object

\item \code{scanMode}: Get the mode of the scan

\item \code{msPlot}: Create a plot

\item \code{nextScan}: Get the next scan

\item \code{previousScan}: Get previous scan

\item \code{parent}: Get the scan containing the parent ion for this scan

\item \code{children}: Get fragmentation scans from this scan

\item \code{siblings}: Get all scans with the same parent

\item \code{msInfo}: Get information about the scan

\item \code{msData}: Get the scan data
}}
\section{Slots}{

\describe{
\item{\code{connections}}{A list of MsConnections}

\item{\code{info}}{A data.frame with information on the contained scans}

\item{\code{data}}{A matrix containing the raw data pertaining to the scans}

\item{\code{mapping}}{A matrix with the mapping of scans to the rows in the
@data matrix}
}}
\seealso{
Other MSsary-classees: \code{\link{MsChromList-class}},
  \code{\link{[,MsChromList,logical,missing,missing-method}},
  \code{\link{[,MsChromList,numeric,missing,missing-method}},
  \code{\link{msData,MsChromList-method}},
  \code{\link{msInfo,MsChromList-method}},
  \code{\link{msPlot,MsChromList-method}},
  \code{\link{names,MsChromList-method}},
  \code{\link{show,MsChromList-method}};
  \code{\link{MsIonList-class}},
  \code{\link{msData,MsIonList-method}},
  \code{\link{msInfo,MsIonList-method}},
  \code{\link{msPlot,MsIonList-method}},
  \code{\link{names,MsIonList-method}},
  \code{\link{show,MsIonList-method}};
  \code{\link{MsPeakList-class}},
  \code{\link{[,MsPeakList,logical,missing,missing-method}},
  \code{\link{[,MsPeakList,numeric,missing,missing-method}},
  \code{\link{msData,MsPeakList-method}},
  \code{\link{msInfo,MsPeakList-method}},
  \code{\link{msPlot,MsPeakList-method}},
  \code{\link{names,MsPeakList-method}},
  \code{\link{show,MsPeakList-method}}
}

