% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MsScanList-navigators}
\alias{MsScanList-navigators}
\alias{children}
\alias{nextScan}
\alias{parent}
\alias{previousScan}
\alias{siblings}
\title{Methods for navigating scans}
\usage{
nextScan(object, sameLevel = TRUE)

previousScan(object, sameLevel = TRUE)

parent(object)

children(object)

siblings(object)
}
\arguments{
\item{object}{An MsScanList object}

\item{sameLevel}{Should only scans of the same level as the starting scan be
considered?}
}
\value{
Either an MsScanList object or a list of these
}
\description{
These methods facilitates creating new MsScanList objects from an already
defined one. They utilise the inherent relationship between scans in LC-MS/MS
data.
}
\details{
\code{nextScan}: Get the next consecutive scan

\code{previousScan}: Get the previous consecutive scan

\code{parent}: Get the parent scan of a fragmentation scan

\code{children}: Get fragmentation scans from a parent scan

\code{siblings}: Get all fragmentation scans from a fragmentation scans
parent
}
\seealso{
\code{\linkS4class{MsScanList}}
}

