% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsSet.R
\docType{class}
\name{MsSet-class}
\alias{MsSet}
\alias{MsSet-class}
\title{Class to handle MsSet connection}
\description{
This class mainly handles continuous sqlite connection and I/O
}
\section{Methods}{

\describe{
\item{\code{addData(name, data)}}{Adds data to an already created table. 'name' specifies the name of the
table, 'data' is a data.frame containing the data to be added.}

\item{\code{addTable(name, definition, rtree = FALSE)}}{Creates the table 'name' with the columns defined in 'definition'.
The definition can be a character vector that will be concatenated with ', '.}

\item{\code{saryset()}}{Get the saryset database connection in a safe manner}

\item{\code{setData(name, data, key)}}{Insert or update data depending on the existence of key}

\item{\code{updateData(name, data, key)}}{Updates column in already existing data}
}}

