% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsChromList.R
\docType{methods}
\name{peaks,MsChromList-method}
\alias{peaks,MsChromList-method}
\title{Extract peaks from the EIC}
\usage{
\S4method{peaks}{MsChromList}(object, overlaps = FALSE, ...)
}
\description{
Extract peaks from the EIC
}

