% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsData.R
\docType{methods}
\name{peaks,MsData-method}
\alias{peaks,MsData-method}
\title{Extrack peaks from an MsData object}
\usage{
\S4method{peaks}{MsData}(object, ...)
}
\description{
Extrack peaks from an MsData object
}

