% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsIonList.R
\docType{methods}
\name{peaks,MsIonList-method}
\alias{peaks,MsIonList-method}
\title{Extract peaks from an MsIonList}
\usage{
\S4method{peaks}{MsIonList}(object, overlap = FALSE)
}
\description{
Extract peaks from an MsIonList
}

