% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsScanList.R
\docType{methods}
\name{peaks,MsScanList-method}
\alias{peaks,MsScanList-method}
\title{Get peaks intersecting with scan}
\usage{
\S4method{peaks}{MsScanList}(object, parent = FALSE, ...)
}
\description{
Get peaks intersecting with scan
}

