% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepareMethods.R
\docType{class}
\name{prepareMethodStore-class}
\alias{prepareMethodStore}
\alias{prepareMethodStore-class}
\title{Object to handle all peak detection methods}
\description{
This peakMethods object is the sole instance of a peakMethodStore reference
class. Its responsibility is to index all registered peak detection
algorithms and perform input and output check upon calling the functions.
This object should not be interacted with directly but through detectPeaks
methods for the different object from MSsary.
}
\section{Fields}{

\describe{
\item{\code{functions}}{A named list with all registered peak detection functions}

\item{\code{requirements}}{A named list with requirements for all reqistered
functions}

\item{\code{package}}{A named list with the package from where the function was
registered}

\item{\code{version}}{A named list with the version of the package stated in the
package field}
}}
\section{Methods}{

\describe{
\item{\code{registerMethod(name, fun, req = list())}}{Register a new function with the given 'name' and optionally sets
the requirements}

\item{\code{testInput(name, scans, info)}}{Test the provided inputs to see if they are sound and corresponds
to the requirements set when the function was reqistered}

\item{\code{useMethod(name, scans, info, ...)}}{Runs the function given by 'name' with the supplied arguments.
Input and output are automatically tested to ensure conformance.}
}}

