% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsData.R
\docType{methods}
\name{prepareScans,MsData-method}
\alias{prepareScans,MsData-method}
\title{Scan preprocessing in MsData objects}
\usage{
\S4method{prepareScans}{MsData}(object, method, ...)
}
\description{
Scan preprocessing in MsData objects
}

