% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MsIonList.R
\docType{methods}
\name{scans,MsIonList-method}
\alias{scans,MsIonList-method}
\title{Extract the scans that make up an ion list}
\usage{
\S4method{scans}{MsIonList}(object, ...)
}
\description{
Extract the scans that make up an ion list
}

