% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{chroms,MsData-method}
\alias{chroms,MsData-method}
\title{Extract chromatograms from an MsData object}
\usage{
\S4method{chroms}{MsData}(object, ..., mz, raw = FALSE)
}
\description{
This function is used to get chromatographic data from an MsData object. If
nothing besides the object is passed the full chromatogram is extracted, but
this can be altered by passing in scan, mz and retention time constraints.
The returned MsScanList object contains both TIC and BPC so there is no need
to specify this during creation
}

