% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createMsDataSet}
\alias{createMsDataSet}
\title{Create an MsDataSet from a list of raw files}
\usage{
createMsDataSet(rawFiles, setName, memberNames, ref, design, force = FALSE)
}
\description{
Create an MsDataSet from a list of raw files
}

