% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dropEmpty}
\alias{dropEmpty}
\title{Subsets an object to remove empty elements}
\usage{
dropEmpty(object)
}
\arguments{
\item{object}{A list-like object from MSsary}
}
\value{
An object of the same class as the one passed to the method
}
\description{
This function removes empty elements from a list-like MSsary object.
}

