% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getContAcqNum}
\alias{getContAcqNum}
\title{Extract a continuous range of acquisitionNums from the same MS level}
\usage{
getContAcqNum(con, seqNum, retentionTime, msLevel, raw = FALSE)
}
\description{
This function works as getAcqNum, but are constrained to only extract a
single continuous interval from the same MS level such as needed when
extracting peaks etc.
}

