% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MsPeakList-class}
\alias{MsPeakList-class}
\title{Class to handle detected peaks}
\description{
This class is a container for one or several peaks from MsData object(s).
}
\section{Slots}{

\describe{
\item{\code{connections}}{A list of MsConnections}

\item{\code{info}}{A data.frame with information on the contained peaks}

\item{\code{data}}{A matrix containing the raw data pertaining to the peaks}

\item{\code{mapping}}{A matrix with the mapping of peaks to the rows in the
@data matrix}
}}
\seealso{
Other MSsary-classees: \code{\link{MsChromList-class}},
  \code{\link{[,MsChromList,logical,missing,missing-method}},
  \code{\link{[,MsChromList,numeric,missing,missing-method}},
  \code{\link{msData,MsChromList-method}},
  \code{\link{msInfo,MsChromList-method}},
  \code{\link{msPlot,MsChromList-method}},
  \code{\link{show,MsChromList-method}};
  \code{\link{MsIonList-class}},
  \code{\link{[,MsIonList,logical,missing,missing-method}},
  \code{\link{[,MsIonList,numeric,missing,missing-method}},
  \code{\link{length,MsIonList-method}},
  \code{\link{msData,MsIonList-method}},
  \code{\link{msInfo,MsIonList-method}},
  \code{\link{msPlot,MsIonList-method}},
  \code{\link{show,MsIonList-method}};
  \code{\link{MsScanList-class}},
  \code{\link{[,MsScanList,logical,missing,missing-method}},
  \code{\link{[,MsScanList,numeric,missing,missing-method}},
  \code{\link{children,MsScanList-method}},
  \code{\link{msData,MsScanList-method}},
  \code{\link{msInfo,MsScanList-method}},
  \code{\link{msPlot,MsScanList-method}},
  \code{\link{nextScan,MsScanList-method}},
  \code{\link{parent,MsScanList-method}},
  \code{\link{previousScan,MsScanList-method}},
  \code{\link{scanMode,MsScanList-method}},
  \code{\link{show,MsScanList-method}},
  \code{\link{siblings,MsScanList-method}}
}

