% Generated by roxygen2 (4.0.2): do not edit by hand
\name{msPlot}
\alias{msPlot}
\title{Generic plot function for MSsary objects}
\usage{
msPlot(object, ...)
}
\arguments{
\item{object}{An object of an MSsary class}

\item{...}{additional parameters passed along}
}
\value{
The data.frame used to construct the ggplot
}
\description{
This function provides the general plotting interface in MSsary. All MSsary
objects will have this methods, often with several different outputs
depending on other parameters
}

