% Generated by roxygen2 (4.0.2): do not edit by hand
\name{peaks}
\alias{peaks}
\title{Extract peaks from an object}
\usage{
peaks(object, ...)
}
\arguments{
\item{object}{The object to extract ions from}

\item{...}{Aditional parameters depending on the class of @object}
}
\value{
An MsPeakList object
}
\description{
This function allows the user to extract peaks from a range of
objects into an MsPeakList object
}

