% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{prepareScans,MsData-method}
\alias{prepareScans,MsData-method}
\title{Scan preprocessing in MsData objects}
\usage{
\S4method{prepareScans}{MsData}(object, method, ...)
}
\description{
Scan preprocessing in MsData objects
}

