% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{scanMode}
\alias{scanMode}
\alias{scanMode<-}
\title{Get and set the mode of scans in an MsScanList}
\usage{
scanMode(object)

scanMode(object) <- value
}
\arguments{
\item{object}{An MsScanList object}

\item{value}{A vector with 'centroid' or 'profile' values}
}
\value{
An MsScanList
}
\description{
This function returns or sets the mode of the scans in an MsScanList, either
'centroid' or 'profile'. The mode is internally calculated using a heuristic
that looks at the m/z difference between consecutive ions.
}

