% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{scans,MsChromList-method}
\alias{scans,MsChromList-method}
\title{Extract the scans that make up a chromatogram}
\usage{
\S4method{scans}{MsChromList}(object, ...)
}
\description{
Extract the scans that make up a chromatogram
}

