% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{scans,MsData-method}
\alias{scans,MsData-method}
\title{Extract scans from an MsData object}
\usage{
\S4method{scans}{MsData}(object, ...)
}
\description{
This method returns one or several scans based on a range of selection
criteria. See the arguments section for a full list.
}

