% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{scans}
\alias{scans}
\title{Extract scans from an object}
\usage{
scans(object, ...)
}
\arguments{
\item{object}{The object to extract scans from}

\item{...}{Aditional parameters depending on the class of @object}
}
\value{
An MsScanList object
}
\description{
This function allows the user to extract scans from a range of objects into
an MsScanList object
}

