% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{uNames}
\alias{uNames}
\title{Get unique names of elements}
\usage{
uNames(object)
}
\arguments{
\item{object}{A list-like object from MSsary}
}
\value{
A character vector of the same length as object
}
\description{
This function ensures unique element names even in the case of duplicated
elements
}

